requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.2.1
django-dump-load-utf8==0.0.4
djangorestframework==3.16.0
filelock==3.16.1
pillow==11.2.1
platformdirs==4.3.8
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.26.6
